Response types changed in EventController class.

+++ src/main/java/hu/preznyak/eventmngr/controller/event/EventController.java
@@ -5,12 +5,11 @@ import hu.preznyak.eventmngr.model.entity.Event;
 import hu.preznyak.eventmngr.model.repository.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.NoSuchElementException;
-
 @RestController
-@RequestMapping("/")
+@RequestMapping("/event")
 public class EventController {
 
     EventRepository eventRepository;
@@ -20,47 +19,53 @@ public class EventController {
         this.eventRepository = eventRepository;
     }
 
-    @PostMapping("event/create/{title}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Event createEventWithTitle(@PathVariable String title) {
+    @PostMapping("/create/title")
+    public ResponseEntity<Event> createEventWithTitle(@RequestParam("title") String title) {
 
-        return eventRepository.save(new EventBuilder().setTitle(title).createEvent());
+        return new ResponseEntity<>(eventRepository.save(new EventBuilder().setTitle(title).createEvent()), HttpStatus.CREATED);
     }
 
-    @GetMapping("event/findAll")
-    @ResponseStatus(HttpStatus.FOUND)
-    public Iterable<Event> findAllEvent() {
-        return eventRepository.findAll();
+    @GetMapping("/findAll")
+    public ResponseEntity<Iterable<Event>> findAllEvent() {
+        return new ResponseEntity<>(eventRepository.findAll(), HttpStatus.OK);
     }
 
-    @PostMapping("event/create")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Event createEvent(@RequestBody Event event) {
-        return eventRepository.save(new EventBuilder()
+    @PostMapping("/create")
+    public ResponseEntity<Event> createEvent(@RequestBody Event event) {
+        Event newEvent = eventRepository.save(new EventBuilder()
                 .setTitle(event.getTitle()).setDescription(event.getDescription()).setLocation(event.getLocation())
                 .setStartDate(event.getStartDate()).setEndDate(event.getEndDate()).createEvent());
+        return new ResponseEntity<>(newEvent, HttpStatus.CREATED);
     }
 
-    @GetMapping("event/find/{id}")
-    @ResponseStatus(HttpStatus.FOUND)
-    public Event findEventById(@PathVariable Integer id){
-        return eventRepository.findById(id).orElse(null);
+    @GetMapping("/find/{id}")
+    public ResponseEntity<Event> findEventById(@PathVariable Integer id) {
+        Event event = eventRepository.findById(id).orElse(null);
+        if (event == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(event, HttpStatus.OK);
     }
 
-    @DeleteMapping("event/delete/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteEventById(@PathVariable Integer id) {
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<String> deleteEventById(@PathVariable Integer id) {
+        if (eventRepository.findById(id).orElse(null) == null) {
+            return new ResponseEntity<>("Event does not exist.", HttpStatus.BAD_REQUEST);
+        }
         eventRepository.deleteById(id);
+        return new ResponseEntity<>("Event deleted.", HttpStatus.OK);
     }
 
-    @PutMapping("event/update")
-    @ResponseStatus(HttpStatus.OK)
-    public Event updateEvent(@RequestBody Event event) {
-        //TODO Remove the get() method
-        Event toUpdate = eventRepository.findById(event.getId()).get();
+    @PutMapping("/update")
+    public ResponseEntity<Event> updateEvent(@RequestBody Event event) {
+
+        Event toUpdate = eventRepository.findById(event.getId()).orElse(null);
+        if(toUpdate == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         toUpdate.setLocation(event.getLocation());
         toUpdate.setStartDate(event.getStartDate());
         toUpdate.setEndDate(event.getEndDate());
-        return eventRepository.save(toUpdate);
+        return new ResponseEntity<>(eventRepository.save(toUpdate), HttpStatus.OK);
     }
 }
